+++ src/models/posts/validators.rs
@@ -2,11 +2,6 @@ use crate::errors::ValidationError;
 use crate::utils::html::clean_html_strict;
 
 pub const ATTACHMENTS_MAX_NUM: usize = 15;
-pub const EMOJI_MAX_SIZE: usize = 250 * 1000; // 250 kB
-pub const EMOJI_MEDIA_TYPES: [&str; 2] = [
-    "image/gif",
-    "image/png",
-];
 pub const EMOJIS_MAX_NUM: usize = 20;
 
 pub const CONTENT_MAX_SIZE: usize = 100000;