+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Security
+
+- Validate emoji name before saving.
+
 ## [1.10.0] - 2023-01-18
 
 ### Added